feat: react router updates. remove some changes
