+++ test/screenshot/temporary-package/temporary-package.scss
@@ -4,7 +4,3 @@
 body {
   color: blue;
 }
-
-a {
-  color: orange;
-}
